+++ submission/submission-core/src/main/kotlin/ac/uk/ebi/biostd/submission/config/StatsConfig.kt
@@ -1,4 +1,4 @@
-package ac.uk.ebi.biostd.common.config
+package ac.uk.ebi.biostd.submission.config
 
 import ac.uk.ebi.biostd.persistence.common.service.StatsDataService
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceQueryService
