+++ submission/submission-webapp/src/itest/kotlin/ac/uk/ebi/biostd/itest/test/stats/SubmissionStatsTest.kt
@@ -2,7 +2,7 @@ package ac.uk.ebi.biostd.itest.test.stats
 
 import ac.uk.ebi.biostd.client.integration.commons.SubmissionFormat.TSV
 import ac.uk.ebi.biostd.client.integration.web.BioWebClient
-import ac.uk.ebi.biostd.common.config.FilePersistenceConfig
+import ac.uk.ebi.biostd.submission.config.FilePersistenceConfig
 import ac.uk.ebi.biostd.itest.common.SecurityTestService
 import ac.uk.ebi.biostd.itest.entities.RegularUser
 import ac.uk.ebi.biostd.itest.entities.SuperUser
