+++ submission/submission-webapp/src/itest/kotlin/ac/uk/ebi/biostd/itest/test/user/UserGroupsApiTest.kt
@@ -2,7 +2,7 @@ package ac.uk.ebi.biostd.itest.test.user
 
 import ac.uk.ebi.biostd.client.exception.WebClientException
 import ac.uk.ebi.biostd.client.integration.web.BioWebClient
-import ac.uk.ebi.biostd.common.config.FilePersistenceConfig
+import ac.uk.ebi.biostd.submission.config.FilePersistenceConfig
 import ac.uk.ebi.biostd.itest.common.SecurityTestService
 import ac.uk.ebi.biostd.itest.entities.TestUser
 import ac.uk.ebi.biostd.itest.itest.getWebClient
