+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/files/FileConfig.kt
@@ -1,6 +1,5 @@
 package ac.uk.ebi.biostd.files
 
-import ac.uk.ebi.biostd.common.properties.FilesProperties
 import ac.uk.ebi.biostd.files.service.FileServiceFactory
 import ac.uk.ebi.biostd.files.web.common.FilesMapper
 import ebi.ac.uk.ftp.FtpClient
@@ -15,15 +14,4 @@ class FileConfig {
 
     @Bean
     fun fileServiceFactory(ftpClient: FtpClient) = FileServiceFactory(ftpClient)
-
-    companion object {
-        fun ftpClient(fileProperties: FilesProperties): FtpClient {
-            return FtpClient.create(
-                ftpUser = fileProperties.ftpUser,
-                ftpPassword = fileProperties.ftpPassword,
-                ftpUrl = fileProperties.ftpUrl,
-                ftpPort = fileProperties.ftpPort,
-            )
-        }
-    }
 }
